Fix the dtype of the matrix.

+++ src/kde/utils/distanceMatrix.py
@@ -16,5 +16,5 @@ def _compute_distance_matrix_Python(patterns):
 
 def _compute_distance_matrix_C(patterns):
     (num_patterns, _) = patterns.shape
-    distances = np.empty([num_patterns, num_patterns])
+    distances = np.empty([num_patterns, num_patterns], dtype=np.float64)
     return _utils.distance_matrix(patterns, distances)
